requirements.txt
streamlit==1.31.1
pymongo==4.5.0
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
requests==2.31.0
pytest==7.4.3
